requirements.txt
discord.py==2.1.0
requests==2.28.1
beautifulsoup4==4.11.1
